backend/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.4

# LLM & AI
anthropic==0.18.1
openai==1.12.0
tiktoken==0.6.0

# ── ARIA RAG: Embedding model (Phase 1) ───────────────────────────────────────
# all-MiniLM-L6-v2 — CPU-only, zero API cost, Apache 2.0
# torch CPU build keeps the Docker image lean (~500MB vs ~2GB for CUDA)
sentence-transformers==3.0.1
torch==2.3.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
# ─────────────────────────────────────────────────────────────────────────────

# OCR
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── NEW: Admin 2FA ─────────────────────────────────────────────────────────────
pyotp==2.9.0
qrcode[pil]==7.4.2
# ──────────────────────────────────────────────────────────────────────────────

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0
email-validator==2.1.0
